Fix icon imports
While `import Github from './svgs/github.svg'` works for testing in Jest,
this caused issues with icons not rendering on the frontend and triggers
a fatal error.

Once the `.svg` mock was added, the correct syntax from the SVGR docs
`import {ReactComponent as Github} from './svgs/github.svg'` works on the
frontend and in Jest.